pyproject.toml
[project]
name = "enterprise-agent-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-classic>=1.0.3",
    "langchain-google-genai>=4.2.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.12",
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.2.2",
]
